Make parsing more robust in the face of unknown or unexpected tags
Unfortunately it is common for certain tags to be present in the wrong
header, so there are many packages in the wild which rpm-rs would fail
to parse. Additionally any packages built with brand-new versions of
rpmbuild which contain tags not listed in our list of constants would
fail.

In either case, we ought to be a bit more tolerant. Users should be able
to parse the RPMs and do basic things with them even if the new
functionality isn't available.

closes #119

+++ src/tests.rs
@@ -196,7 +196,7 @@ fn test_rpm_header_base(package: RPMPackage) -> Result<(), Box<dyn std::error::E
         let actual_entry = &metadata.signature.index_entries[i];
         assert_eq!(*len as u32, actual_entry.num_items);
         assert_eq!(*data, actual_entry.data);
-        assert_eq!(*tag, actual_entry.tag);
+        assert_eq!(*tag as u32, actual_entry.tag);
     }
 
     assert_eq!(
